Moved and organized quest patches
